+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 - Cantidad maxima de materias Seguimiento Academico
+- Impresion de Comprobantes de Pago
 
 ## [2.4.0] - 2023-06-14
 ### Added